Refactored Stream request into common lib
